Consolidate validation utils and rename qc modules
- Rename utils/qc.py -> utils/validation.py
- Rename quality_control/qc_binaries -> qc_booleans (and test file)
- Move BarycentricCorrection._strictly_increasing into
  utils/validation.strictly_increasing (generalized to a numeric array)
- Use strictly_increasing in RadialVelocity.compute_ccf to guard that each
  order's wavelengths are strictly ascending (also rejects constant/NaN rows)
