--- kpfpipe/quality_control/qc_binaries/__init__.py
@@ -1,6 +1,0 @@
-from kpfpipe.quality_control.qc_binaries.base import QC
-from kpfpipe.quality_control.qc_binaries.level0 import QCL0
-from kpfpipe.quality_control.qc_binaries.level1 import QCL1
-from kpfpipe.quality_control.qc_binaries.level2 import QCL2
-
-__all__ = ["QC", "QCL0", "QCL1", "QCL2"]
