+++ recipes/kpf_drp_science.py
@@ -12,7 +12,7 @@ from kpfpipe.modules.barycentric_correction import BarycentricCorrection
 from kpfpipe.modules.radial_velocity import RadialVelocity
 
 from kpfpipe.quality_control.diagnostics import DiagL1, DiagL2
-from kpfpipe.quality_control.qc_binaries import QCL1, QCL2
+from kpfpipe.quality_control.qc_booleans import QCL1, QCL2
 from kpfpipe.quality_control.quicklook.level0 import PlotL0
 from kpfpipe.quality_control.quicklook.level1 import PlotL1
